repository interface 로 분리
